ScottPlot4: improve Palette base classes
